rename : 설정 정보를 config 폴더 내에서 관리하도록 변경
