+++ src/common/interceptor/success.interceptor.ts
@@ -1,0 +1,19 @@
+import {
+  CallHandler,
+  ExecutionContext,
+  Injectable,
+  NestInterceptor,
+} from '@nestjs/common';
+import { map, Observable } from 'rxjs';
+
+@Injectable()
+export class SuccessInterceptor implements NestInterceptor {
+  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
+    return next.handle().pipe(
+      map((data) => {
+        const res = context.switchToHttp().getResponse();
+        res.json(data);
+      }),
+    );
+  }
+}
