+++ src/config/interceptorConfig.ts
@@ -1,0 +1,5 @@
+import { SuccessInterceptor } from '../common/interceptor/success.interceptor';
+
+export const interceptorConfig = (app) => {
+  app.useGlobalPipes(new SuccessInterceptor());
+};
