+++ src/config/pipeConfig.ts
@@ -1,0 +1,5 @@
+import { validationPipe } from '../common/pipe/validation.pipe';
+
+export const pipeConfig = (app) => {
+  validationPipe(app);
+};
