+++ src/main.ts
@@ -1,14 +1,17 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { validationPipe } from './common/pipe/validation.pipe';
-import { sessionConfig } from './module/auth/auth.session.config';
+import { sessionConfig } from './config/session.config';
+import { interceptorConfig } from './config/interceptorConfig';
+import { pipeConfig } from './config/pipeConfig';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   app.setGlobalPrefix('/api');
-  validationPipe(app);
+  pipeConfig(app);
 
   sessionConfig(app);
+
+  interceptorConfig(app);
 
   const port = process.env.PORT;
   await app.listen(port);
